feat: expand sort methods for SparseVector refactor: introduce operations module for SparseVector
