+++ graphblas_sparse_linear_algebra/src/collections/sparse_vector/mod.rs
@@ -2,6 +2,8 @@
 mod element;
 mod sparse_vector;
 
+pub mod operations;
+
 pub use element::{VectorElement, VectorElementList};
 pub use sparse_vector::{
     FromVectorElementList, GetVectorElement, GetVectorElementList, GetVectorElementValue,
