+++ graphblas_sparse_linear_algebra/src/collections/sparse_vector/operations/mod.rs
@@ -1,0 +1,4 @@
+// mod sort_matrix;
+mod sort;
+
+pub use sort::SortSparseVector;
